split the main file into a few making it easier to read
