+++ bin/lumen/src/error.rs
@@ -1,0 +1,13 @@
+use lumen_rollkit::PayloadAttributesError;
+use thiserror::Error;
+
+/// Custom error type used in payload attributes validation
+#[derive(Debug, Error)]
+pub enum RollkitEngineError {
+    #[error("Invalid transaction data: {0}")]
+    InvalidTransactionData(String),
+    #[error("Gas limit exceeded")]
+    GasLimitExceeded,
+    #[error("Rollkit payload attributes error: {0}")]
+    PayloadAttributes(#[from] PayloadAttributesError),
+}
