+++ bin/lumen/src/main.rs
@@ -5,181 +5,52 @@
 
 #![allow(missing_docs, rustdoc::missing_crate_level_docs)]
 
-use alloy_eips::{eip2718::Decodable2718, eip4895::Withdrawals};
-use alloy_primitives::{Address, Bytes, B256, U256};
-use alloy_rpc_types::{
-    engine::{
-        ExecutionData, ExecutionPayloadEnvelopeV2, ExecutionPayloadEnvelopeV3,
-        ExecutionPayloadEnvelopeV4, ExecutionPayloadEnvelopeV5, ExecutionPayloadV1,
-        PayloadAttributes as EthPayloadAttributes, PayloadId,
-    },
-    Withdrawal,
+pub mod attributes;
+pub mod builder;
+pub mod error;
+
+use alloy_rpc_types::engine::{
+    ExecutionData, ExecutionPayloadEnvelopeV2, ExecutionPayloadEnvelopeV3,
+    ExecutionPayloadEnvelopeV4, ExecutionPayloadEnvelopeV5, ExecutionPayloadV1,
 };
 use clap::Parser;
-use lumen_node::{RollkitPayloadBuilder, RollkitPayloadBuilderConfig};
-use lumen_rollkit::{PayloadAttributesError, RollkitPayloadAttributes};
-use reth_basic_payload_builder::{
-    BuildArguments, BuildOutcome, HeaderForPayload, PayloadBuilder, PayloadConfig,
-};
-use reth_engine_local::payload::UnsupportedLocalAttributes;
 use reth_ethereum::{
-    chainspec::{ChainSpec, ChainSpecProvider},
+    chainspec::ChainSpec,
     node::{
         api::{
             payload::{EngineApiMessageVersion, EngineObjectValidationError, PayloadOrAttributes},
             validate_version_specific_fields, AddOnsContext, EngineTypes, EngineValidator,
             FullNodeComponents, FullNodeTypes, InvalidPayloadAttributesError, NewPayloadError,
-            NodeTypes, PayloadAttributes, PayloadBuilderAttributes, PayloadTypes, PayloadValidator,
+            NodeTypes, PayloadTypes, PayloadValidator,
         },
         builder::{
-            components::{BasicPayloadServiceBuilder, ComponentsBuilder, PayloadBuilderBuilder},
+            components::{BasicPayloadServiceBuilder, ComponentsBuilder},
             rpc::{EngineValidatorBuilder, RpcAddOns},
-            BuilderContext, Node, NodeAdapter, NodeComponentsBuilder,
+            Node, NodeAdapter, NodeComponentsBuilder,
         },
         node::{
             EthereumConsensusBuilder, EthereumExecutorBuilder, EthereumNetworkBuilder,
             EthereumPoolBuilder,
         },
-        EthEvmConfig, EthereumEthApiBuilder,
+        EthereumEthApiBuilder,
     },
-    pool::{PoolTransaction, TransactionPool},
-    primitives::{Header, RecoveredBlock, SealedBlock},
-    TransactionSigned,
+    primitives::{RecoveredBlock, SealedBlock},
 };
 use reth_ethereum_cli::{chainspec::EthereumChainSpecParser, Cli};
 use reth_ethereum_payload_builder::EthereumExecutionPayloadValidator;
-use reth_payload_builder::{EthBuiltPayload, EthPayloadBuilderAttributes, PayloadBuilderError};
-use reth_provider::HeaderProvider;
-use reth_revm::cached::CachedReads;
+use reth_payload_builder::EthBuiltPayload;
 use reth_trie_db::MerklePatriciaTrie;
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
-use thiserror::Error;
 use tracing::info;
+
+use crate::{
+    attributes::{RollkitEnginePayloadAttributes, RollkitEnginePayloadBuilderAttributes},
+    builder::{RollkitArgs, RollkitPayloadBuilderBuilder},
+};
 
 #[global_allocator]
 static ALLOC: reth_cli_util::allocator::Allocator = reth_cli_util::allocator::new_allocator();
-
-/// Rollkit-specific command line arguments
-#[derive(Debug, Clone, Parser, PartialEq, Eq, Serialize, Deserialize, Default)]
-pub struct RollkitArgs {
-    /// Enable Rollkit mode for the node (enabled by default)
-    #[arg(
-        long = "rollkit.enable",
-        default_value = "true",
-        help = "Enable Rollkit integration for transaction processing via Engine API"
-    )]
-    pub enable_rollkit: bool,
-}
-
-/// Rollkit payload attributes that support passing transactions via Engine API
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RollkitEnginePayloadAttributes {
-    /// Standard Ethereum payload attributes
-    #[serde(flatten)]
-    pub inner: EthPayloadAttributes,
-    /// Transactions to be included in the payload (passed via Engine API)
-    pub transactions: Option<Vec<Bytes>>,
-    /// Optional gas limit for the payload
-    #[serde(rename = "gasLimit")]
-    pub gas_limit: Option<u64>,
-}
-
-impl UnsupportedLocalAttributes for RollkitEnginePayloadAttributes {}
-
-/// Custom error type used in payload attributes validation
-#[derive(Debug, Error)]
-pub enum RollkitEngineError {
-    #[error("Invalid transaction data: {0}")]
-    InvalidTransactionData(String),
-    #[error("Gas limit exceeded")]
-    GasLimitExceeded,
-    #[error("Rollkit payload attributes error: {0}")]
-    PayloadAttributes(#[from] PayloadAttributesError),
-}
-
-impl PayloadAttributes for RollkitEnginePayloadAttributes {
-    fn timestamp(&self) -> u64 {
-        self.inner.timestamp()
-    }
-
-    fn withdrawals(&self) -> Option<&Vec<Withdrawal>> {
-        self.inner.withdrawals()
-    }
-
-    fn parent_beacon_block_root(&self) -> Option<B256> {
-        self.inner.parent_beacon_block_root()
-    }
-}
-
-/// Rollkit payload builder attributes
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct RollkitEnginePayloadBuilderAttributes {
-    /// Ethereum payload builder attributes
-    pub ethereum_attributes: EthPayloadBuilderAttributes,
-    /// Decoded transactions from the Engine API
-    pub transactions: Vec<TransactionSigned>,
-    /// Gas limit for the payload
-    pub gas_limit: Option<u64>,
-}
-
-impl PayloadBuilderAttributes for RollkitEnginePayloadBuilderAttributes {
-    type RpcPayloadAttributes = RollkitEnginePayloadAttributes;
-    type Error = RollkitEngineError;
-
-    fn try_new(
-        parent: B256,
-        attributes: RollkitEnginePayloadAttributes,
-        _version: u8,
-    ) -> Result<Self, Self::Error> {
-        let ethereum_attributes = EthPayloadBuilderAttributes::new(parent, attributes.inner);
-
-        // Decode transactions from bytes if provided
-        let mut transactions = Vec::new();
-        if let Some(tx_bytes_vec) = attributes.transactions {
-            for tx_bytes in tx_bytes_vec {
-                // Decode the transaction from RLP bytes (as sent by Go client)
-                let tx = TransactionSigned::network_decode(&mut tx_bytes.as_ref())
-                    .map_err(|e| RollkitEngineError::InvalidTransactionData(e.to_string()))?;
-                transactions.push(tx);
-            }
-        }
-
-        Ok(Self {
-            ethereum_attributes,
-            transactions,
-            gas_limit: attributes.gas_limit,
-        })
-    }
-
-    fn payload_id(&self) -> PayloadId {
-        self.ethereum_attributes.id
-    }
-
-    fn parent(&self) -> B256 {
-        self.ethereum_attributes.parent
-    }
-
-    fn timestamp(&self) -> u64 {
-        self.ethereum_attributes.timestamp
-    }
-
-    fn parent_beacon_block_root(&self) -> Option<B256> {
-        self.ethereum_attributes.parent_beacon_block_root
-    }
-
-    fn suggested_fee_recipient(&self) -> Address {
-        self.ethereum_attributes.suggested_fee_recipient
-    }
-
-    fn prev_randao(&self) -> B256 {
-        self.ethereum_attributes.prev_randao
-    }
-
-    fn withdrawals(&self) -> &Withdrawals {
-        &self.ethereum_attributes.withdrawals
-    }
-}
 
 /// Rollkit engine types - uses custom payload attributes that support transactions
 #[derive(Clone, Debug, Default, Deserialize, Serialize)]
@@ -410,200 +281,6 @@ where
 
     fn add_ons(&self) -> Self::AddOns {
         RollkitNodeAddOns::default()
-    }
-}
-
-/// Rollkit payload service builder that integrates with the rollkit payload builder
-#[derive(Debug, Clone)]
-#[non_exhaustive]
-pub struct RollkitPayloadBuilderBuilder {
-    config: RollkitPayloadBuilderConfig,
-}
-
-impl RollkitPayloadBuilderBuilder {
-    /// Create a new builder with rollkit args
-    pub fn new(_args: &RollkitArgs) -> Self {
-        let config = RollkitPayloadBuilderConfig {
-            max_transactions: 1000,
-            min_gas_price: 1_000_000_000, // 1 Gwei
-        };
-        info!("Created Rollkit payload builder with config: {:?}", config);
-        Self { config }
-    }
-}
-
-impl Default for RollkitPayloadBuilderBuilder {
-    fn default() -> Self {
-        Self::new(&RollkitArgs::default())
-    }
-}
-
-/// The rollkit engine payload builder that integrates with the rollkit payload builder
-#[derive(Debug, Clone)]
-pub struct RollkitEnginePayloadBuilder<Pool, Client>
-where
-    Pool: Clone,
-    Client: Clone,
-{
-    rollkit_builder: Arc<RollkitPayloadBuilder<Client>>,
-    #[allow(dead_code)]
-    pool: Pool,
-    #[allow(dead_code)]
-    config: RollkitPayloadBuilderConfig,
-}
-
-impl<Pool, Client> PayloadBuilder for RollkitEnginePayloadBuilder<Pool, Client>
-where
-    Client: reth_ethereum::provider::StateProviderFactory
-        + ChainSpecProvider<ChainSpec = ChainSpec>
-        + HeaderProvider<Header = Header>
-        + Clone
-        + Send
-        + Sync
-        + 'static,
-    Pool: TransactionPool<Transaction: PoolTransaction<Consensus = TransactionSigned>>,
-{
-    type Attributes = RollkitEnginePayloadBuilderAttributes;
-    type BuiltPayload = EthBuiltPayload;
-
-    fn try_build(
-        &self,
-        args: BuildArguments<Self::Attributes, Self::BuiltPayload>,
-    ) -> Result<BuildOutcome<Self::BuiltPayload>, PayloadBuilderError> {
-        let BuildArguments {
-            cached_reads: _,
-            config,
-            cancel: _,
-            best_payload,
-        } = args;
-        let PayloadConfig {
-            parent_header,
-            attributes,
-        } = config;
-
-        info!(
-            "Rollkit engine payload builder: building payload with {} transactions",
-            attributes.transactions.len()
-        );
-
-        // Convert Engine API attributes to Rollkit payload attributes
-        let rollkit_attrs = RollkitPayloadAttributes::new(
-            attributes.transactions.clone(),
-            attributes.gas_limit,
-            attributes.timestamp(),
-            attributes.prev_randao(),
-            attributes.suggested_fee_recipient(),
-            attributes.parent(),
-            parent_header.number + 1,
-        );
-
-        // Build the payload using the rollkit payload builder - use spawn_blocking for async work
-        let rollkit_builder = self.rollkit_builder.clone();
-        let sealed_block = tokio::task::block_in_place(|| {
-            tokio::runtime::Handle::current().block_on(rollkit_builder.build_payload(rollkit_attrs))
-        })
-        .map_err(PayloadBuilderError::other)?;
-
-        info!(
-            "Rollkit engine payload builder: built block with {} transactions, gas used: {}",
-            sealed_block.transaction_count(),
-            sealed_block.gas_used
-        );
-
-        // Convert to EthBuiltPayload
-        let gas_used = sealed_block.gas_used;
-        let built_payload = EthBuiltPayload::new(
-            attributes.payload_id(), // Use the proper payload ID from attributes
-            Arc::new(sealed_block),
-            U256::from(gas_used), // Block gas used
-            None,                 // No blob sidecar for rollkit
-        );
-
-        if let Some(best) = best_payload {
-            if built_payload.fees() <= best.fees() {
-                return Ok(BuildOutcome::Aborted {
-                    fees: built_payload.fees(),
-                    cached_reads: CachedReads::default(),
-                });
-            }
-        }
-
-        Ok(BuildOutcome::Better {
-            payload: built_payload,
-            cached_reads: CachedReads::default(),
-        })
-    }
-
-    fn build_empty_payload(
-        &self,
-        config: PayloadConfig<Self::Attributes, HeaderForPayload<Self::BuiltPayload>>,
-    ) -> Result<Self::BuiltPayload, PayloadBuilderError> {
-        let PayloadConfig {
-            parent_header,
-            attributes,
-        } = config;
-
-        info!("Rollkit engine payload builder: building empty payload");
-
-        // Create empty rollkit attributes (no transactions)
-        let rollkit_attrs = RollkitPayloadAttributes::new(
-            vec![],
-            attributes.gas_limit,
-            attributes.timestamp(),
-            attributes.prev_randao(),
-            attributes.suggested_fee_recipient(),
-            attributes.parent(),
-            parent_header.number + 1,
-        );
-
-        // Build empty payload - use spawn_blocking for async work
-        let rollkit_builder = self.rollkit_builder.clone();
-        let sealed_block = tokio::task::block_in_place(|| {
-            tokio::runtime::Handle::current().block_on(rollkit_builder.build_payload(rollkit_attrs))
-        })
-        .map_err(PayloadBuilderError::other)?;
-
-        let gas_used = sealed_block.gas_used;
-        Ok(EthBuiltPayload::new(
-            attributes.payload_id(), //TODO: make sure this works
-            Arc::new(sealed_block),
-            U256::from(gas_used),
-            None,
-        ))
-    }
-}
-
-impl<Node, Pool> PayloadBuilderBuilder<Node, Pool, EthEvmConfig> for RollkitPayloadBuilderBuilder
-where
-    Node: FullNodeTypes<
-        Types: NodeTypes<
-            Payload = RollkitEngineTypes,
-            ChainSpec = ChainSpec,
-            Primitives = reth_ethereum::EthPrimitives,
-        >,
-    >,
-    Pool: TransactionPool<Transaction: PoolTransaction<Consensus = TransactionSigned>>
-        + Unpin
-        + 'static,
-{
-    type PayloadBuilder = RollkitEnginePayloadBuilder<Pool, Node::Provider>;
-
-    async fn build_payload_builder(
-        self,
-        ctx: &BuilderContext<Node>,
-        pool: Pool,
-        evm_config: EthEvmConfig,
-    ) -> eyre::Result<Self::PayloadBuilder> {
-        let rollkit_builder = Arc::new(RollkitPayloadBuilder::new(
-            Arc::new(ctx.provider().clone()),
-            evm_config,
-        ));
-
-        Ok(RollkitEnginePayloadBuilder {
-            rollkit_builder,
-            pool,
-            config: self.config,
-        })
     }
 }
